Update FreePlotDigitizer.m
Changed the docstring to something shorter. I'll this description to warn that the user specified segments change I made just before this is based on linear interpolation, so it won't work with other graph types. The next thing I'll do is add this for curve mode, and also make it work with other axis types as well.